Changed link from Youtube channels to Youtube playlist
The Laravel Daily is a very useful channel for anyone looking to learn and deepen their knowledge of Laravel and stay up to date with the latest developments in the Laravel world. Daily 5-10 minute videos help me a lot in learning Laravel! This playlist demonstrates how to build a Laravel website from the start and from the perspective of an experienced Laravel programmer.